requirements.txt
anyio==3.7.0
atproto==0.0.12
bases==0.2.1
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dacite==1.8.1
dag-cbor==0.3.2
Flask==2.3.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
httpx-ws==0.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multiformats==0.2.1
multiformats-config==0.2.0.post4
peewee==3.16.2
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.0
typing-validation==1.0.0.post2
typing_extensions==4.5.0
Werkzeug==2.3.4
wsproto==1.2.0
